chore: rename new files to kebab case
