+++ src/files/file-preview/file-thumbnail.css
