+++ src/icons/stroke-icons.js
